# [closes #12304] marginally faster dictionary creation.