--- a/packages/ember-metal/lib/dictionary.js
+++ b/packages/ember-metal/lib/dictionary.js
@@ -1,3 +1,4 @@
+import EmptyObject from './empty_object';
 
 // the delete is meant to hint at runtimes that this object should remain in
 // dictionary mode. This is clearly a runtime specific hack, but currently it
@@ -5,7 +6,14 @@
 // the cost of creation dramatically over a plain Object.create. And as this
 // only makes sense for long-lived dictionaries that aren't instantiated often.
 export default function makeDictionary(parent) {
-  var dict = Object.create(parent);
+  var dict;
+
+  if (arguments.length === 0 || parent === null) {
+    dict = new EmptyObject();
+  } else {
+    dict = Object.create(parent);
+  }
+
   dict['_dict'] = null;
   delete dict['_dict'];
   return dict;