on chrome 47:

```
makeDictionary(null) ..................... 2,390,190.74 op/s
makeDictionaryWithEmptyObject ............ 3,377,259.20 op/s
```

---

Safari: Version 8.0.7 (10600.7.12)
```
makeDictionary(null) ..................... 1,606,191.49 op/s
makeDictionaryWithEmptyObject ............ 1,595,042.30 op/s
```

---

Saf Version 8.0.7 (10600.7.12, r189464)

```
makeDictionary(null) ..................... 2,460,767.99 op/s
makeDictionaryWithEmptyObject ............ 3,182,898.12 op/s
```

---

FF 43.0a1 (2015-08-18)
```
makeDictionary(null) ..................... 8,872,924.67 op/s
makeDictionaryWithEmptyObject ............ 8,665,248.18 op/s
```