@stefanpenner 2015-09-07
I wonder if the container/registry structure, should just be EmptyViews not in dict mode..