@rwjblue 2015-09-07
@stefanpenner - Definitely agreed. AFAICT from looking at the code, `delete`'s only happen when calling `registry.unregister` which should be much less often.