@rwjblue 2015-09-07
Looks like one failure (snippet from Travis logs):

```
Test failed: Ember.Application Dependency Injection - default resolver: knownForType returns each item for a given type found
    Failed assertion: Died on test #1     at http://localhost:13141/qunit/qunit.js:263
    at http://localhost:13141/ember-tests.js:2207
    at internalRequire (http://localhost:13141/ember.debug.js:78)
    at http://localhost:13141/ember.debug.js:44
    at http://localhost:13141/tests/?package=ember-application:173: No default value
TypeError: No default value
```