@homu 2016-01-06
:umbrella: The latest upstream changes (presumably #12754) made this pull request unmergeable. Please resolve the merge conflicts.