@stefanpenner 2016-01-06
https://github.com/emberjs/ember.js/pull/12754